Type Checking!
Been looking into the level format for Minecraft New Nintendo 3DS Edition, and I think this is the database implementation that it uses. I'm not sure whether this JS version itself fully works with the file I have, but I thought I may as well update things and modernize it just in case it is what I need. Knowing how the library works helps me understand how I use it itself. Contributing is always a bonus as well! :D

About the level format exploration:

Not sure if this is the same format, but I also have hope that it just may be
I think I might have discovered this last time we looked into it, but I thought it wasn't the right one for some reason
https://en.wikipedia.org/wiki/Cdb_(software)