TSC Migration
Not sure how to default export from JSDoc-powered JS, when in CommonJS. Once I move to ESM and or TS it's not an issue, but I was kind of surprised that this didn't work already, since it's an existing practice for Node packages that have been around for a while. I'm probably just doing it wrong, but I'm not totally sure. I thought it might've been the part where it was assigning both the variable and the `module.exports` in the same line, but that didn't fix the errors either.

https://github.com/microsoft/TypeScript/issues/37832
https://stackoverflow.com/questions/60009092/declaration-will-not-emit-due-to-private-name-usage
https://github.com/microsoft/TypeScript/issues/2719

Super bowl!!!!! aaaeaaeggh, 380 green hut, AA